Pulled out Unsafe/ScheduledTask.
The primary reason is that there was significant performance overhead on
accessing the localTaskQueue through the Unsafe because it wasn't a
field on Unsafe, but on the enclosing class (the Eventloop). And I guess
the JIT ran into problems because Eventloop itself is abstract.

So now it is clear; there is no more hidden field access to enclosing
classes.
