+++ hazelcast-tpc-engine/src/main/java/com/hazelcast/internal/tpc/nio/NioEventloop.java
@@ -19,6 +19,7 @@ package com.hazelcast.internal.tpc.nio;
 import com.hazelcast.internal.tpc.AsyncServerSocket;
 import com.hazelcast.internal.tpc.AsyncSocket;
 import com.hazelcast.internal.tpc.Eventloop;
+import com.hazelcast.internal.tpc.Unsafe;
 import com.hazelcast.internal.tpc.util.NanoClock;
 
 import java.io.IOException;
@@ -56,7 +57,7 @@ public final class NioEventloop extends Eventloop {
 
     @Override
     protected Unsafe createUnsafe() {
-        return new NioUnsafe();
+        return new NioUnsafe(this);
     }
 
     @Override
@@ -138,6 +139,4 @@ public final class NioEventloop extends Eventloop {
         closeQuietly(selector);
     }
 
-    private class NioUnsafe extends Unsafe {
-    }
 }
